refactor(theme): separate resolver contracts and RGB
